[Chat] Split Local bridge into Cache and Memory bridges
