+++ examples/commands/message-stores.php
@@ -13,11 +13,11 @@ require_once dirname(__DIR__).'/bootstrap.php';
 
 use Doctrine\DBAL\DriverManager;
 use MongoDB\Client as MongoDbClient;
+use Symfony\AI\Chat\Bridge\Cache\Store as CacheStore;
 use Symfony\AI\Chat\Bridge\Doctrine\DoctrineDbalMessageStore;
 use Symfony\AI\Chat\Bridge\HttpFoundation\SessionStore;
-use Symfony\AI\Chat\Bridge\Local\CacheStore;
-use Symfony\AI\Chat\Bridge\Local\InMemoryStore;
 use Symfony\AI\Chat\Bridge\Meilisearch\MessageStore as MeilisearchMessageStore;
+use Symfony\AI\Chat\Bridge\Memory\Store as InMemoryStore;
 use Symfony\AI\Chat\Bridge\MongoDb\MessageStore as MongoDbMessageStore;
 use Symfony\AI\Chat\Bridge\Pogocache\MessageStore as PogocacheMessageStore;
 use Symfony\AI\Chat\Bridge\Redis\MessageStore as RedisMessageStore;
