+++ src/chat/src/Bridge/Memory/Store.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\AI\Chat\Bridge\Local;
+namespace Symfony\AI\Chat\Bridge\Memory;
 
 use Symfony\AI\Chat\ManagedStoreInterface;
 use Symfony\AI\Chat\MessageStoreInterface;
@@ -18,7 +18,7 @@ use Symfony\AI\Platform\Message\MessageBag;
 /**
  * @author Christopher Hertel <mail@christopher-hertel.de>
  */
-final class InMemoryStore implements ManagedStoreInterface, MessageStoreInterface
+final class Store implements ManagedStoreInterface, MessageStoreInterface
 {
     /**
      * @var MessageBag[]
